Feat: "Completed ROLLUP for Subtotals Exercise"
This commit includes the exercises completed from the 'Intermediate SQL for Data Scientists' course on LinkedIn Learning, specifically the section on using ROLLUP to create subtotals. These exercises have enhanced my understanding of how to use the ROLLUP function in SQL for creating subtotals, which is a valuable skill for data analysis. The knowledge and skills gained from these exercises will be beneficial in handling and analyzing data in future data science projects.